Usernames are now case insensitive.
Fixes bug #11

This fixes the issue with usernames being case sensitive to the queue logic, but not SQL or LDAP.
Now usernames are fully case insensitve on the server side.
Additionally, this fix didn't require changes to the client side javascript, and should not affect the dozen or so users that have logged in with upper or mixed case.
Finally, all new users will be stored in SQL with lowercase usernames.

ZZ